% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_distance.r
\name{abc_distance}
\alias{abc_distance}
\title{abc_distance}
\usage{
abc_distance(type_sum, summaries, summ_weights, weight)
}
\arguments{
\item{type_sum}{choose between classic, model based and mixed}

\item{summaries}{vector of summaries}

\item{summ_weights}{vector with weights of the summaries computed from a pilot study}

\item{weight}{weights for the model-based summaries}
}
\value{
ABC distance for the given summaries
}
\description{
Compute the distance given the summaries
}
