% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_sequential_varying_threshold_number_sim.R
\name{abc_sequential_varying_threshold_number_sim}
\alias{abc_sequential_varying_threshold_number_sim}
\title{abc_sequential_varying_threshold_number_sim}
\usage{
abc_sequential_varying_threshold_number_sim(
  data,
  extra,
  extra_summaries,
  parmask,
  ABCthreshold,
  number_sim,
  summ_weights,
  numparticles,
  alpha,
  sampling,
  attempt,
  nu,
  type_copula,
  type_margin,
  nu_marg,
  folder,
  we,
  whichsummarymodelbased,
  whichprior = 1,
  subsamplingby = 1,
  whichiter = 5
)
}
\arguments{
\item{data}{observed data}

\item{extra}{Any additional coefficients/vector of coefficients which may be need to simulate from the model}

\item{extra_summaries}{Any additional coefficients/vector needed to compute the summaries}

\item{parmask}{set of 0/1 to see which parameters have to be estimated}

\item{ABCthreshold}{Initial value for the threshold for the guided approach}

\item{number_sim}{Total number of simulations, after which we stop the algorithm}

\item{summ_weights}{vector with weights of the summaries computed from a pilot study}

\item{numparticles}{number of particles to keep}

\item{alpha}{percentile to compute the ABC threshold out of all distances}

\item{sampling}{specify the chosen samplers. The options are blocked, blockedopt, standard, olcm, full_cond, full_cond_opt}

\item{attempt}{number of iteration}

\item{nu}{degree of freedom of the t Copula (so type_copula =2)}

\item{type_copula}{0 for the multivariate Gaussian, 1 for the Gaussian copula, 2 for the t-copula}

\item{type_margin}{1 for triangular, 2 for t scaled, 3 for logistic, 5 for uniform, 6 for gaussian}

\item{nu_marg}{degree of freedom of the t location-scale marginal (so type_marginal =2)}

\item{folder}{specify the folder where the results are going to be saved in}

\item{we}{weights to scale the model-based summaries}

\item{whichsummarymodelbased}{specify whether to consider IAEspectrum, IAE density, Wass density, Wass spectrum}

\item{whichprior}{1 for uniform, 2 for lognormal priors}

\item{subsamplingby}{every how many simulated points the observation should be taken. The default=1, i.e., no subsampling}

\item{whichiter}{Specify from when which iteration you want the new sampler to be used}
}
\description{
Guided and non-guided SIS-ABC and SMC-ABC for time-varying thresholds computed as percentiles of ACCEPTED distances with a fixed budget in terms of number of simulations
}
