% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_summaries.r
\name{abc_summaries}
\alias{abc_summaries}
\title{abc_summaries}
\usage{
abc_summaries(x, extra_summaries, whichsummarymodelbased)
}
\arguments{
\item{x}{Matrix with observations;}

\item{extra_summaries}{List with any extra coefficients which may be needed in the summaries}

\item{whichsummarymodelbased}{specify whether to consider IAEspectrum, IAE density, Wass density, Wass spectrum}
}
\value{
Invariant spectral density and invariant density
}
\description{
Summaries
}
