% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_summaries.r
\name{abc_summaries_classic}
\alias{abc_summaries_classic}
\title{abc_summaries_classic}
\usage{
abc_summaries_classic(x)
}
\arguments{
\item{x}{Matrix with observations;}
}
\value{
Summaries
}
\description{
Summaries
}
