% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_summaries.r
\name{FHN_abc_summaries_X}
\alias{FHN_abc_summaries_X}
\title{FHN_abc_summaries_X}
\usage{
FHN_abc_summaries_X(X, span_val, Lsupport)
}
\arguments{
\item{X}{V component of the observed data}

\item{span_val}{span value for the smoothing periodogram}

\item{Lsupport}{double check this}
}
\value{
List with step size for the invariant spectral density and invariant density, values of the spectral density and invariant density
}
\description{
Summaries for the reference data X, namely spectral density and invariant density.
}
