% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.r
\name{model}
\alias{model}
\title{model}
\usage{
model(theta, extra)
}
\arguments{
\item{theta}{underlying parameter}

\item{extra}{Vector with time step delta, initial condition X0 and number of points N to simulate}
}
\value{
Simulation of observations from all model, returned as matrix;
}
\description{
Code for simulating using the splitting scheme from the FHN model
}
