% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_theta.r
\name{sample_theta}
\alias{sample_theta}
\title{sample_theta}
\usage{
sample_theta(
  proposal_mean,
  proposal_cov,
  nu,
  type_copula,
  type_margin,
  nu_marg
)
}
\arguments{
\item{proposal_mean}{mean of the sampler}

\item{proposal_cov}{covariance matrix of the sampler}

\item{nu}{degree of freedom in case type_copula =2 (t student)}

\item{type_copula}{0 for the multivariate Gaussian, 1 for the Gaussian copula, 2 for the t-copula, 3 for the Frank copula for 2-dim parameter vector}

\item{type_margin}{1 for triangular, 2 for t scaled, 3 for logistic, 4 for lognormal, 5 for uniform, 6 for gaussian}

\item{nu_marg}{degree of freedom of the t location-scale marginal (so type_marginal =2)}
}
\value{
d-dimensional sampled value
}
\description{
Simulate 1 sample for theta
}
