% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.r
\name{model}
\alias{model}
\title{model}
\usage{
model(theta, extra)
}
\arguments{
\item{theta}{underlying parameter}

\item{extra}{Any additional coefficients/vector of coefficients which may be need to simulate from the model}
}
\value{
Simulation of observations from all model, returned as matrix;
}
\description{
Considered models
}
