% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMCABC_numbersim.R
\name{SMCABC_numbersim}
\alias{SMCABC_numbersim}
\title{SMCABC_numbersim}
\usage{
SMCABC_numbersim(
  data,
  extra,
  extra_summaries,
  ABCthreshold,
  number_sim,
  summ_weights,
  numparticles,
  alpha,
  sampling,
  attempt,
  folder,
  we,
  whichsummarymodelbased,
  whichprior = "unif",
  subsamplingby = 1
)
}
\arguments{
\item{data}{observed data}

\item{extra}{Vector with time step delta, initial condition X0 and number of points N to simulate}

\item{extra_summaries}{Any additional coefficients/vector needed to compute the summaries}

\item{ABCthreshold}{Initial value for the threshold for the guided approach}

\item{number_sim}{Total number of simulations, after which we stop the algorithm}

\item{summ_weights}{vector with weights of the summaries computed from a pilot study}

\item{numparticles}{number of particles to keep}

\item{alpha}{percentile to compute the ABC threshold out of all distances}

\item{sampling}{specify the chosen samplers. The options are standard or olcm}

\item{attempt}{number of iteration}

\item{folder}{specify the folder where the results are going to be saved in}

\item{we}{weights to scale the structure-based summaries}

\item{whichsummarymodelbased}{specify whether to consider IAEspectrum, IAE density, Wass density, Wass spectrum}

\item{whichprior}{choose between 'unif','lognormal' and and 'exp'}

\item{subsamplingby}{every how many simulated points the observation should be taken. The default=1, i.e., no subsampling}
}
\description{
SMC-ABC for time-varying thresholds computed as percentiles of ACCEPTED distances with a fixed budget (number of simulations)
}
